requirements.txt
accelerate==1.5.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
airportsdata==20250224
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
astor==0.8.1
attrs==25.2.0
audioread==3.0.1
av==14.2.0
blake3==1.0.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
codetiming==1.4.0
compressed-tensors==0.9.1
contourpy==1.3.1
cupy-cuda12x==13.4.0
cycler==0.12.1
datasets==3.3.2
decorator==5.2.1
deepspeed==0.16.4
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
docstring-parser==0.16
editorconfig==0.17.0
einops==0.8.1
email-validator==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
fastrlock==0.8.3
ffmpy==0.5.0
filelock==3.17.0
fire==0.7.0
flash-attn @ https://github.com/Dao-AILab/flash-attention/releases/download/v2.7.3/flash_attn-2.7.3+cu12torch2.5cxx11abiFALSE-cp312-cp312-linux_x86_64.whl
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
gguf==0.10.0
gitdb==4.0.12
gitpython==3.1.44
gradio==5.20.1
gradio-client==1.7.2
groovy==0.1.2
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.3
hydra-core==1.3.2
idna==3.10
importlib-metadata==8.6.1
iniconfig==2.0.0
interegular==0.3.3
jieba==0.42.1
jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsbeautifier==1.15.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
-e file:///home/tlin4/projects/LangGFM
lark==1.2.2
lazy-loader==0.4
librosa==0.11.0
liger-kernel==0.5.4
littleutils==0.2.4
llvmlite==0.43.0
lm-format-enforcer==0.10.11
markdown-it-py==3.0.0
markupsafe==2.1.5
mathruler==0.1.0
matplotlib==3.10.1
mdurl==0.1.2
mistral-common==1.5.3
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
ogb==1.3.6
omegaconf==2.3.0
openai==1.66.2
opencv-python-headless==4.11.0.86
orjson==3.10.15
outdated==0.2.2
outlines==0.1.11
outlines-core==0.1.26
packaging==24.2
pandas==2.2.3
partial-json-parser==0.2.1.1.post5
peft==0.14.0
pgeocode==0.5.0
pillow==11.1.0
pip==25.0.1
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.0.2
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pybind11==2.13.6
pycountry==24.6.1
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydub==0.25.1
pygments==2.19.1
pylatexenc==2.10
pyparsing==3.2.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
pyyaml==6.0.2
pyzmq==26.3.0
qwen-vl-utils==0.0.10
ray==2.40.0
rdkit==2024.9.6
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rouge-chinese==1.0.3
rpds-py==0.23.1
ruff==0.9.10
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.22.0
setproctitle==1.3.5
setuptools==76.0.0
shellingham==1.5.4
shtab==1.7.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
sse-starlette==2.2.1
starlette==0.46.1
sympy==1.13.1
tabulate==0.9.0
tensordict==0.6.2
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.5.1
torch-geometric==2.6.1
torchaudio==2.5.1
torchdata==0.11.0
torchvision==0.20.1
tqdm==4.67.1
transformers==4.49.0
triton==3.1.0
trl==0.15.2
typeguard==4.4.2
typer==0.15.2
typing-extensions==4.12.2
tyro==0.9.16
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vllm==0.7.3
wandb==0.19.8
watchfiles==1.0.4
websockets==15.0.1
xformers==0.0.28.post3
xgrammar==0.1.11
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
